refactor: updated ui with the refactored Ui states
