+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/navigation/DukanNavHost.kt
@@ -6,12 +6,12 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.toRoute
+import net.thechance.mena.dukan.presentation.screen.categoryDukans.DukansScreen
 import net.thechance.mena.dukan.presentation.screen.createDukan.CreateDukanScreen
 import net.thechance.mena.dukan.presentation.screen.createProduct.CreateProductScreen
 import net.thechance.mena.dukan.presentation.screen.createShelf.CreateShelfScreen
 import net.thechance.mena.dukan.presentation.screen.dukanCategories.DukanCategoriesScreen
 import net.thechance.mena.dukan.presentation.screen.dukanDetails.DukanDetailsScreen
-import net.thechance.mena.dukan.presentation.screen.dukans.DukansScreen
 import net.thechance.mena.dukan.presentation.screen.main.MainScreen
 import net.thechance.mena.dukan.presentation.screen.manageDukan.ManageDukanScreen
 import net.thechance.mena.dukan.presentation.screen.manageShelf.ManageShelfScreen
