+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/createProduct/component/ShelfSection.kt
@@ -19,7 +19,7 @@ import net.thechance.mena.designsystem.presentation.component.chip.Chip
 import net.thechance.mena.designsystem.presentation.component.icon.Icon
 import net.thechance.mena.designsystem.presentation.component.text.Text
 import net.thechance.mena.designsystem.presentation.theme.theme.Theme
-import net.thechance.mena.dukan.presentation.viewModel.createProduct.ShelfUiState
+import net.thechance.mena.dukan.presentation.viewModel.createProduct.CreateProductUiState.ShelfUiState
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
