+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/util/stubPreviews/PreviewCreateDukanInteractionListener.kt
@@ -3,10 +3,10 @@ package net.thechance.mena.dukan.presentation.util.stubPreviews
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.unit.DpOffset
 import com.attafitamim.krop.core.images.ImageSrc
-import net.thechance.mena.dukan.presentation.viewModel.createDukan.ColorUiState
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanInteractionListener
 import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState
-import net.thechance.mena.dukan.presentation.viewModel.createDukan.DukanCategoryUiState
+import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState.ColorUiState
+import net.thechance.mena.dukan.presentation.viewModel.createDukan.CreateDukanUiState.DukanCategoryUiState
 import org.maplibre.compose.camera.CameraPosition
 
 object PreviewCreateDukanInteractionListener : CreateDukanInteractionListener {
