+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/util/stubPreviews/PreviewManageDukanInteractionListener.kt
@@ -2,8 +2,8 @@ package net.thechance.mena.dukan.presentation.util.stubPreviews
 
 import net.thechance.mena.dukan.presentation.component.SnackBarType
 import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ManageDukanInteractionListener
-import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ProductUiState
-import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ShelfUiState
+import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ManageDukanUiState.ProductUiState
+import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ManageDukanUiState.ShelfUiState
 import org.jetbrains.compose.resources.StringResource
 
 object PreviewManageDukanInteractionListener : ManageDukanInteractionListener {
